Rename `litesvm` plugin to `litesvmConnection`
This PR renames the `litesvm` plugin to `litesvmConnection` to free up the `litesvm` name for a future all-in-one plugin, matching the pattern established by the RPC plugins. All source files, tests, docblocks, and READMEs are updated accordingly.
