+++ .changeset/salty-donuts-shout.md
@@ -1,0 +1,6 @@
+---
+'@solana/kit-plugin-litesvm': minor
+'@solana/kit-client-litesvm': patch
+---
+
+Rename `litesvm` plugin to `litesvmConnection`.
