+++ packages/kit-plugin-litesvm/src/index.browser.ts
@@ -1,10 +1,10 @@
 export type { FailedTransactionMetadata, TransactionMetadata } from 'litesvm';
-export type { LiteSVM } from './litesvm';
+export type { LiteSVM } from './litesvm-connection';
 export type { LiteSvmRpcApi } from './litesvm-to-rpc';
 
-export function litesvm(): <T extends object>(_client: T) => never {
+export function litesvmConnection(): <T extends object>(_client: T) => never {
     throw new Error(
-        'The `litesvm` plugin is unavailable in browser and react-native. ' +
+        'The `litesvmConnection` plugin is unavailable in browser and react-native. ' +
             'Use this plugin in a node environment instead.',
     );
 }
