+++ packages/kit-plugin-litesvm/src/index.ts
@@ -2,7 +2,7 @@ export type { FailedTransactionMetadata, TransactionMetadata } from 'litesvm';
 
 export * from './airdrop';
 export * from './get-minimum-balance';
-export * from './litesvm';
+export * from './litesvm-connection';
 export * from './litesvm-to-rpc';
 export * from './transaction-error';
 export * from './transaction-plan-executor';
