+++ packages/kit-plugin-litesvm/test/litesvm-connection.test.ts
@@ -13,14 +13,14 @@ import {
 import { LiteSVM } from 'litesvm';
 import { describe, expect, expectTypeOf, it } from 'vitest';
 
-import { litesvm as nodeLitesvm } from '../src/index';
-import { litesvm as browserLitesvm } from '../src/index.browser';
+import { litesvmConnection as nodeLitesvm } from '../src/index';
+import { litesvmConnection as browserLitesvm } from '../src/index.browser';
 const litesvm = __NODEJS__ ? nodeLitesvm : browserLitesvm;
 
 describe('litesvm', () => {
     if (!__NODEJS__) {
         it('throws in browser builds', () => {
-            expect(litesvm).toThrow('The `litesvm` plugin is unavailable in browser and react-native');
+            expect(litesvm).toThrow('The `litesvmConnection` plugin is unavailable in browser and react-native');
         });
         return;
     }
