+++ packages/kit-plugin-litesvm/test/litesvm-to-rpc.test.ts
@@ -1,14 +1,14 @@
 import { address, createClient, generateKeyPairSigner, lamports, Rpc } from '@solana/kit';
 import { describe, expect, it } from 'vitest';
 
-import { litesvm as nodeLitesvm } from '../src/index';
-import { litesvm as browserLitesvm } from '../src/index.browser';
+import { litesvmConnection as nodeLitesvm } from '../src/index';
+import { litesvmConnection as browserLitesvm } from '../src/index.browser';
 const litesvm = __NODEJS__ ? nodeLitesvm : browserLitesvm;
 
 describe('createRpcFromSvm', () => {
     if (!__NODEJS__) {
         it('throws in browser builds', () => {
-            expect(litesvm).toThrow('The `litesvm` plugin is unavailable in browser and react-native');
+            expect(litesvm).toThrow('The `litesvmConnection` plugin is unavailable in browser and react-native');
         });
         return;
     }
